refactor(Function): 함수별로 파일을 분리하여 사용
